Made use of RevenuePerDate instead of Map<LocalDate, Double>, also added growthPercentage in the revenue by month

+++ src/main/java/com/baseshelf/analytics/RevenueMonthDto.java
@@ -11,5 +11,5 @@ public class RevenueMonthDto {
     private String month; // jan to dec
     private Integer monthNumber; // 1 to 12
     private Double revenue; // revenue of that month
-    private Float growthPercentage; //profit made compare to previous month
+    private Double growthPercentage; //profit made compare to previous month
 }
